requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchainhub
langchain-google-genai 
asyncpg
streamlit
PyPDF2
python-dotenv
requests
pypdf
python-socketio
langchain-anthropic
duckduckgo-search
midjourney-api
langchain-chroma
faiss-cpu
chromadb
langchain-ollama
langserve
fastapi
uvicorn
sse_starlette
langchain-tavily
streamlit-lottie
youtube-transcript-api
google-search-results

tavily-python

pydantic_ai
langgraph

Ipython
docling